cleanup vulkan utils
